Refactor Rule to store list of Symbols
This is preferable to using strings because it will allow us to decorate
each symbol with additional attributes.

This refactor is not complete, tests are broken for now and there are
more opportunities to replace strings with symbols, but its a good start
and the main project does build and run without (compile-time) errors.